requirements.txt
asgiref==3.2.7
Django==3.0.5
django-el-pagination==3.3.0
django-jquery==3.1.0
pytz==2019.3
sqlparse==0.3.1
coverage==4.5.4
python-coveralls==2.9.3
django-filter==2.2.0
django-bootstrap-form==3.4
